+++ Dates.ts
@@ -53,7 +53,7 @@ module Dates {
      * (i.e.if {@code incrementAtMidnight == true } then 2001-3-15 2:43 is the same day as 2001-3-15 19:39, even though the dates are more than 12 hours apart)
      */
     export function dayDiff(dtLeft: Date, dtRight: Date, incrementAtMidnight: boolean = false): number {
-        var daysDiff = ((<number>dtLeft.getTime() - <number>dtRight.getTime()) / DateConstants.MS_PER_DAY);
+        var daysDiff = ((<number>dtLeft.getTime() - <number>dtRight.getTime()) / DateConstants.MILLIS_PER_DAY);
         var dateDiff = (incrementAtMidnight ? Math.floor(daysDiff) : Math.round(daysDiff)); // TODO this does not handle leap years or non-gregorian calendar days
         return dateDiff;
     }