+++ test/DateTimesTest.js
@@ -19,7 +19,7 @@ suite("DateTimes", function DateTimesTest() {
         asr.equal(now, res);
     }
     test("DateTimes.MS_PER_DAY", function MS_PER_DAYTest() {
-        asr.equal(DateConstants.MS_PER_DAY, 24 * 60 * 60 * 1000);
+        asr.equal(DateConstants.MILLIS_PER_DAY, 24 * 60 * 60 * 1000);
     });
     test("DotNetJsonDate.parseDotNetJson", function parseDotNetJsonTest() {
         testDotNetJson(asr);